cvm-rs: Don't pass surface Buffer to JS
This ends up being a big safety issue, especially given the fun fact that the buffer can vanish while JS is still trying to marshal it somewhere else (which happened during testing more than once).

So, displays now no longer own a buffer and instead the VNC engine now does JPEG encoding. This is a bit meh, but it's fine, and at least respects locking the surface and everything from the Rust side (the old code did *not*, which is why it had the risk of crashing), which means it should be MUCH more stable.

Additionally, since JS doesn't have to even perform the operation anymore it should be a bit faster on the main thread too.

The high level JS wrapper to the JPEG encode API is no longer needed in cvmts, nor is sharp, and the cvm-rs JPEG encoder API does not need to be exposed to Javascript anymore, however has been left in/deprecated.