cvmts: Initial PoC implementation of multi-node
As it says on the tin, this changeset refactors the server heavily to allow it to do multinode (run multiple VMs in one server process, basically)

This is very experiemental, so expect bugs/don't run this if you aren't prepared for bugs/etc. (I'm aware of a few, which need to be fixed. Also, once this is merged/fully working, the handbook will need to be altered somewhat to account for it.)